// src/api/generated/model/accreditationStatusEnum.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */

/**
 * * `accredited` - Accredited
 * * `provisional` - Provisional
 * * `pending` - Pending
 * * `expired` - Expired
 * * `not_applicable` - Not Applicable
 */
export type AccreditationStatusEnum = typeof AccreditationStatusEnum[keyof typeof AccreditationStatusEnum];


export const AccreditationStatusEnum = {
  accredited: 'accredited',
  provisional: 'provisional',
  pending: 'pending',
  expired: 'expired',
  not_applicable: 'not_applicable',
} as const;