// src/api/generated/model/facilityTypeEnum.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */

/**
 * * `hospital` - Hospital
 * * `surgery_center` - Ambulatory Surgery Center
 * * `clinic` - Clinic
 * * `physician_office` - Physician Office
 * * `other` - Other
 */
export type FacilityTypeEnum = typeof FacilityTypeEnum[keyof typeof FacilityTypeEnum];


export const FacilityTypeEnum = {
  hospital: 'hospital',
  surgery_center: 'surgery_center',
  clinic: 'clinic',
  physician_office: 'physician_office',
  other: 'other',
} as const;